TFP-5430: Vise frem oppgavebeskrivelser fra Gosys for aksjonspunkter 5033 og 5034 * OppgaveType for kodeverk
